+++ packages/ui/core-components/src/lib/molecules/dimension-grid/DimensionCut.svelte
@@ -8,7 +8,10 @@
 	import { getDimensionCutQuery } from './dimensionGridQuery.js';
 	import { buildQuery } from '@evidence-dev/component-utilities/buildQuery';
 	import getColumnSummary from '@evidence-dev/component-utilities/getColumnSummary';
-	import { formatValue } from '@evidence-dev/component-utilities/formatting';
+	import {
+		formatValue,
+		getFormatObjectFromString
+	} from '@evidence-dev/component-utilities/formatting';
 	import QueryLoad from '../../atoms/query-load/QueryLoad.svelte';
 	import { resolveMaybePromise } from '@evidence-dev/sdk/usql';
 	import Alert from '../../atoms/alert/Alert.svelte';
@@ -19,6 +22,7 @@
 	export let metric;
 	export let metricLabel;
 	export let limit;
+	export let fmt = undefined;
 
 	if (dimension.column_name.includes(' ')) {
 		dimension.column_name = `"${dimension.column_name}"`;
@@ -122,6 +126,7 @@
 		</Alert>
 		{#if loaded?.length > 0 || (Array.isArray(selectedValue) && selectedValue.length > 0)}
 			{@const columnSummary = getColumnSummary(loaded, 'array')?.filter((d) => d.id === 'metric')}
+			{@const fmtObject = fmt ? getFormatObjectFromString(fmt, 'number') : columnSummary[0].format}
 			<div class="transition-all" style={`min-height:${minRem}rem;`}>
 				{#each loaded as row (row.dimensionValue)}
 					<div
@@ -135,9 +140,9 @@
 						<DimensionRow
 							{row}
 							{selectedValue}
-							value={formatValue(
+							formattedValue={formatValue(
 								row.metric,
-								columnSummary[0].format,
+								fmtObject,
 								columnSummary[0].columnUnitSummary
 							)}
 						/>
@@ -157,9 +162,9 @@
 							<DimensionRow
 								row={{ dimensionValue: missingValue }}
 								{selectedValue}
-								value={formatValue(
+								formattedValue={formatValue(
 									missingValue,
-									columnSummary[0].format,
+									fmtObject,
 									columnSummary[0].columnUnitSummary
 								)}
 							/>