+++ packages/ui/core-components/src/lib/unsorted/viz/core/_Chart.svelte
@@ -26,7 +26,6 @@
 	import ErrorChart from './ErrorChart.svelte';
 	import checkInputs from '@evidence-dev/component-utilities/checkInputs';
 	import { chartColours, uiColours } from '@evidence-dev/component-utilities/colours';
-	import EmptyChart from './EmptyChart.svelte';
 
 	// ---------------------------------------------------------------------------------------
 	// Input Props
@@ -1053,28 +1052,24 @@
 
 {#if !error}
 	<slot />
-	{#if !$config.series.length}
-		<EmptyChart emptySet="pass" />
-	{:else}
-		<ECharts
-			config={$config}
-			{height}
-			{width}
-			{data}
-			{queryID}
-			evidenceChartTitle={title}
-			{showAllXAxisLabels}
-			{swapXY}
-			{echartsOptions}
-			{seriesOptions}
-			{printEchartsConfig}
-			{renderer}
-			{downloadableData}
-			{downloadableImage}
-			{connectGroup}
-			{seriesColors}
-		/>
-	{/if}
+	<ECharts
+		config={$config}
+		{height}
+		{width}
+		{data}
+		{queryID}
+		evidenceChartTitle={title}
+		{showAllXAxisLabels}
+		{swapXY}
+		{echartsOptions}
+		{seriesOptions}
+		{printEchartsConfig}
+		{renderer}
+		{downloadableData}
+		{downloadableImage}
+		{connectGroup}
+		{seriesColors}
+	/>
 {:else}
 	<ErrorChart {error} {chartType} />
 {/if}